Initial Commit
Now as we now how to add multiple layers of neurons and make a network of neurons, its time to add some non-linearity in the network. As the real world problems are almost non-linear and we want from our neural network to come up with the function that can map inputs into desired outputs. So in order to bring no linearity in our system we add Activation Functions for each neurons. Activation Functions convert the neuron output to some corresponding output which can be thought of how much the data part is important in prediction.