Initial Commit
Now we have implement the loss function for calculating the loss or error of the network made during training or prediction. This loss functions will be used to optimized the network based on these during training. The Neural Network will learn to minimize the loss in each step and will be optimize. We implemented Categorical Cross-Entropy Loss Function for now, we will implement other loss functions as well as we move forward.